+++ extensions/gif-search/src/preferences.ts
@@ -3,12 +3,13 @@ import fetch from "node-fetch";
 import path from "path";
 import differenceInDays from "date-fns/differenceInDays";
 
-import { getPreferenceValues, environment } from "@raycast/api";
+import { getPreferenceValues, environment, Grid } from "@raycast/api";
 
 export const API_KEY = "apiKey";
 export const MAX_RESULTS = "maxResults";
 export const DEFAULT_ACTION = "defaultAction";
 export const LAYOUT = "layoutType";
+export const GRID_ITEM_SIZE = "gridItemSize";
 
 export const CONFIG_URL = "https://cdn.joe.sh/gif-search/config.json";
 
@@ -109,4 +110,14 @@ export const LAYOUT_TYPE: { [type: string]: LayoutType } = {
 
 export function getLayoutType() {
   return getPrefs()[LAYOUT];
+}
+
+export const GRID_SIZE: { [key: string]: Grid.ItemSize } = {
+  small: Grid.ItemSize.Small,
+  medium: Grid.ItemSize.Medium,
+  large: Grid.ItemSize.Large,
+};
+
+export function getGridItemSize() {
+  return GRID_SIZE[getPrefs()[GRID_ITEM_SIZE]];
 }