+++ extensions/gif-search/CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Added Grid Layout Support] - 2022-06-07
 - Adds new option to display results in a Grid or a List
 - Removes "Show GIF Preview" preference. GIF previews are now solid and should always be shown
+- Adds preference for controlling the Grid Item Size
 
 ## [New Metadata] - 2022-05-09
 - Adds a new icon